Fix bug on maxpool
